Jpk/coupons

+++ src/Payments.Sql/dbo/Tables/LineItems.sql
@@ -3,7 +3,7 @@
     [Amount]      DECIMAL (18, 2) NOT NULL,
     [Description] NVARCHAR (MAX)  NULL,
     [InvoiceId]   INT             NULL,
-    [Quantity]    INT             NOT NULL,
+    [Quantity]    DECIMAL (18, 2) NOT NULL,
     [Total]       DECIMAL (18, 2) NOT NULL,
     CONSTRAINT [PK_LineItems] PRIMARY KEY CLUSTERED ([Id] ASC),
     CONSTRAINT [FK_LineItems_Invoices_InvoiceId] FOREIGN KEY ([InvoiceId]) REFERENCES [dbo].[Invoices] ([Id])
